printer_client/requirements.txt
websockets>=12.0
python-escpos>=3.1
pyusb>=1.2.1
